Update to Xcode 4.6.  Further update to LLVM 3.2.

+++ MLKCons.h
@@ -131,7 +131,7 @@
 
 -(id) copyWithZone:(NSZone *)zone;
 -(BOOL) isEqual:(id)object;
--(unsigned int) hash;
+-(NSUInteger) hash;
 
 -(void) dealloc;
 @end
